pipe line/manga_ai/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.44.0
accelerate>=0.33.0
diffusers>=0.30.0
safetensors>=0.4.3
huggingface_hub>=0.24.0
pydantic>=2.6.0
numpy>=1.26.0,<2.0
Pillow>=10.2.0
opencv-python>=4.9.0.80
scipy>=1.12.0
gradio>=4.44.0
fastapi>=0.110.0
uvicorn>=0.27.0
flask>=3.0.0
segment-anything>=1.0
requests>=2.31.0
